7.2 ver 2

+++ quadratic/form.py
@@ -3,9 +3,9 @@ from django import forms
 
 
 class QuadraticForm(forms.Form):
-    a = forms.FloatField(label="коэффициент а")
-    b = forms.FloatField(label="коэффициент b")
-    c = forms.FloatField(label="коэффициент c")
+    a = forms.IntegerField(label="коэффициент а")
+    b = forms.IntegerField(label="коэффициент b")
+    c = forms.IntegerField(label="коэффициент c")
 
     def clean_a(self):
         a = self.cleaned_data['a']
